Is active middleware

+++ studdit/apps/studdit/src/app/comments/comments.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { User } from '../users/schemas/user.schema';
@@ -15,12 +15,23 @@ export class CommentsService {
         @InjectModel(Thread.name) private threadModel: Model<Thread>,
     ) {}
 
+    private async userIsActive(username: string) {
+        const user = await this.userModel.findOne({ username });
+        return user?.isActive ?? false;
+    } 
+
     async create(createCommentDto: CreateCommentDto): Promise<Comment> {
         try {
             const user = await this.userModel.findOne({ username: createCommentDto.username });
 
             if (!user) {
                 throw new NotFoundException('User not found');
+            }
+
+            const isActive = await this.userIsActive(createCommentDto.username);
+
+            if (!isActive) {
+                throw new UnauthorizedException('User is not active');
             }
 
             const thread = await this.threadModel.findOne({ _id: createCommentDto.threadId });
@@ -47,6 +58,12 @@ export class CommentsService {
                 throw new NotFoundException('User not found');
             }
 
+            const isActive = await this.userIsActive(createNestedCommentDto.username);
+
+            if (!isActive) {
+                throw new UnauthorizedException('User is not active');
+            }
+
             const parentComment = await this.commentModel.findOne({ _id: id });
 
             if (!parentComment) {
@@ -63,12 +80,22 @@ export class CommentsService {
         }
     }
 
-    async delete (id: string): Promise<Comment> {
+    async delete (id: string, username: string): Promise<Comment> {
         try {
             const comment = await this.commentModel.findOne({ _id: id });
 
             if (!comment) {
                 throw new NotFoundException('Comment not found');
+            }
+
+            const isActive = await this.userIsActive(username);
+
+            if (!isActive) {
+                throw new UnauthorizedException('User is not active');
+            }
+
+            if (comment.username !== username) {
+                throw new UnauthorizedException('User is not author of comment');
             }
 
             const deletedComment = await this.commentModel.findByIdAndDelete(id);
@@ -90,6 +117,12 @@ export class CommentsService {
 
             if (!user) {
                 throw new NotFoundException('User not found');
+            }
+
+            const isActive = await this.userIsActive(username);
+
+            if (!isActive) {
+                throw new UnauthorizedException('User is not active');
             }
 
             const hasUpvoted = comment.upvotes.includes(user._id);
@@ -125,6 +158,12 @@ export class CommentsService {
 
             if (!user) {
                 throw new NotFoundException('User not found');
+            }
+
+            const isActive = await this.userIsActive(username);
+
+            if (!isActive) {
+                throw new UnauthorizedException('User is not active');
             }
 
             const hasUpvoted = comment.upvotes.includes(user._id);
